Reconfigure Travis CI for no virtualenv; #1081
Travis CI starts us off in a virtualenv, but we want to own the
virtualenv. I think it's actually a better experience for users if we
don't let them run bootstrap.py in another virtualenv, because it gives
us a chance to communicate Gittip's worldview on the matter (we're using
virtualenv as a library, not a cli). Let's see if we can convince Travis
to go along with us here.